Project renamed, added temperature simulation.
