+++ datacollection/datacollection-demo/source/android/app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
   -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.kaaproject.kaa.demo.smarttrashcan">
+    package="org.kaaproject.kaa.demo.datacollection">
 
   <uses-permission  android:name="android.permission.INTERNET" />
   <uses-permission  android:name="android.permission.ACCESS_NETWORK_STATE" />
@@ -28,7 +28,7 @@
       android:supportsRtl="true"
       android:theme="@style/AppTheme">
     <activity
-        android:name=".MainActivity"
+        android:name="org.kaaproject.kaa.demo.datacollection.MainActivity"
         android:label="@string/app_name">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
