+++ datacollection/datacollection-demo/source/android/app/src/main/java/org/kaaproject/kaa/demo/datacollection/Logger.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kaaproject.kaa.demo.smarttrashcan;
+package org.kaaproject.kaa.demo.datacollection;
 
 /**
  * Project: SmartTrashCan
